completes - Solutions: Extending the API

+++ spring/hplus/src/main/java/com/example/hplus/data/Product.java
@@ -4,6 +4,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "PRODUCTS")
@@ -14,7 +15,8 @@ public class Product {
     String name;
     int size;
     String variety;
-    double price;
+    //double price;
+    BigDecimal price;
     String status;
 
     public String getId() {
@@ -49,11 +51,11 @@ public class Product {
         this.variety = variety;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
